requirements.txt
boto==2.34.0
Fom==0.9.10
GitPython==0.3.2.1
Jinja2==2.7.3
PyPDF2==1.23
arrow==0.4.4
beautifulsoup4==4.3.2
lettuce==0.2.20
requests==2.5.3
wsgiref==0.1.2
lxml==3.4.1
xlrd==0.9.3
PyYAML==3.11
Wand==0.4.0
-e git+https://github.com/elifesciences/jats-scraper.git#egg=jats-scraper
-e hg+https://bitbucket.org/lskibinski/scraper/src#egg=scraper
-e git+https://github.com/elifesciences/elife-tools.git@exp#egg=elifetools
